use a small test data file
