+++ straxen/__init__.py
@@ -26,6 +26,9 @@ from . import analyses
 # Otherwise we have straxen.demo() etc.
 from . import contexts
 
+from . import test_utils
+from test_utils import *
+
 try:
     from . import holoviews_utils
     from .holoviews_utils import *
