+++ straxen/common.py
@@ -2,7 +2,6 @@ import ast
 import configparser
 import gzip
 import inspect
-import io
 import commentjson
 import json
 import os
@@ -11,7 +10,6 @@ import pickle
 import dill
 import socket
 import sys
-import tarfile
 import urllib.request
 import tqdm
 import numpy as np
@@ -55,6 +53,7 @@ LAST_MISCABLED_RUN = 8796
 TSTART_FIRST_CORRECTLY_CABLED_RUN = 1596036001000000000
 
 INFINITY_64BIT_SIGNED = 9223372036854775807
+
 
 @export
 def pmt_positions(xenon1t=False):
@@ -298,16 +297,6 @@ def get_secret(x):
 
 
 @export
-def download_test_data():
-    """Downloads strax test data to strax_test_data in the current directory"""
-    blob = get_resource('https://raw.githubusercontent.com/XENONnT/strax_auxiliary_files/353b2c60a01e96f67e4ba544ce284bd91241964d/strax_files/strax_test_data_straxv1.1.0.tar',  #  noqa
-                        fmt='binary')
-    f = io.BytesIO(blob)
-    tf = tarfile.open(fileobj=f)
-    tf.extractall()
-
-
-@export
 def get_livetime_sec(context, run_id, things=None):
     """Get the livetime of a run in seconds. If it is not in the run metadata,
     estimate it from the data-level metadata of the data things.
@@ -353,26 +342,8 @@ def pre_apply_function(data, run_id, target, function_name='pre_apply_function')
 
 
 def _overwrite_testing_function_file(function_file):
-    """For testing purposes allow this function file to be loaded from HOME/testing_folder"""
-    if not straxen._is_on_pytest():
-        # If we are not on a pytest, never try using a local file.
-        return function_file
-
-    home = os.environ.get('HOME')
-    if home is None:
-        # Impossible to load from non-existent folder
-        return function_file
-
-    testing_file = os.path.join(home, function_file)
-
-    if os.path.exists(testing_file):
-        # For testing purposes allow loading from 'home/testing_folder'
-        warn(f'Using local function: {function_file} from {testing_file}! '
-             f'If you are not integrated testing on github you should '
-             f'absolutely remove this file. (See #559)')
-        function_file = testing_file
-
-    return function_file
+    warn('Use straxen.test_utils._overwrite_testing_function_file')
+    return straxen._overwrite_testing_function_file(function_file)
 
 
 @export
