+++ tests/test_plugins.py
@@ -4,6 +4,7 @@ import numpy as np
 from immutabledict import immutabledict
 import straxen
 from straxen.common import pax_file, aux_repo
+from straxen.test_utils import nt_test_run_id
 ##
 # Tools
 ##
@@ -49,7 +50,7 @@ testing_config_1T = dict(
     electron_drift_time_gate=("electron_drift_time_gate_constant", 1700),
 )
 
-test_run_id_nT = '008900'
+nt_test_run_id = '008900'
 test_run_id_1T = '180423_1021'
 
 
@@ -128,7 +129,7 @@ forbidden_plugins = tuple([p for p in
 
 def _run_plugins(st,
                  make_all=False,
-                 run_id=test_run_id_nT,
+                 run_id=nt_test_run_id,
                  **proces_kwargs):
     """
     Try all plugins (except the DAQReader) for a given context (st) to see if
@@ -276,16 +277,18 @@ def test_nT(ncores=1):
     if ncores == 1:
         print('-- nT lazy mode --')
     init_database = straxen.utilix_is_configured(warning_message=False)
-    st = straxen.contexts.xenonnt_online(_database_init=init_database,
-                                         use_rucio=False)
+    st = straxen.test_utils.nt_test_context(
+        _database_init=init_database,
+        use_rucio=False,
+    )
     offline_gain_model = ("to_pe_placeholder", True)
     _update_context(st, ncores, fallback_gains=offline_gain_model, nt=True)
     # Lets take an abandoned run where we actually have gains for in the CMT
-    _run_plugins(st, make_all=True, max_workers=ncores, run_id=test_run_id_nT)
+    _run_plugins(st, make_all=True, max_workers=ncores, run_id=nt_test_run_id)
     # Test issue #233
     st.search_field('cs1')
     # Test of child plugins:
-    _test_child_options(st, test_run_id_nT)
+    _test_child_options(st, nt_test_run_id)
     print(st.context_config)
